feat: create auth with  jwt

+++ api/src/routes/memories.ts
@@ -3,8 +3,16 @@ import { z } from 'zod';
 import { prisma } from '../lib/prisma';
 
 export async function memoriesRoute(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify();
+    });
+
+    app.get('/memories', async (request) => {
+        console.log(request.user);
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             },
@@ -19,7 +27,7 @@ export async function memoriesRoute(app: FastifyInstance) {
         });
     });
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -31,6 +39,10 @@ export async function memoriesRoute(app: FastifyInstance) {
                 id,
             },
         });
+
+        if (memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         return memory;
     });
@@ -49,19 +61,29 @@ export async function memoriesRoute(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '74f738e2-321c-4e6a-ae2e-8b5a7c70332a',
+                userId: request.user.sub,
             },
         });
 
         return memory;
     });
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(request.params);
+
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        });
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         const bodySchema = z.object({
             content: z.string(),
@@ -71,7 +93,7 @@ export async function memoriesRoute(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-        const memory = await prisma.memory.update({
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -85,12 +107,22 @@ export async function memoriesRoute(app: FastifyInstance) {
         return memory;
     });
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(request.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        });
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         await prisma.memory.delete({
             where: {
